requirements.txt
streamlit
pandas
numpy
scikit-learn==1.6.1
lightgbm
shap
matplotlib
joblib
groq
openai
python-dotenv
langchain
langchain-core
langchain-community
langchain-cerebras
langgraph
pydantic
orjson
scipy
tqdm
watchdog
imbalanced-learn
